refactor(commit): extract write commit pipeline to runtime/commit
The commit pipeline — sendToWriteCh, dispatcher, per-shard processor,
runCommitBurst/Single, syncWorker, ack/finish/release/collect helpers,
applyRequests + writeRequestsToLSM + prepareLSMRequest — moves into
runtime/commit/Pipeline. The Pipeline owns its queue + dispatch fan-out
+ sync worker; the host (DB) provides the rest via commit.Host.

- runtime/commit/pipeline.go (~820 lines): Pipeline, Config, LSM/Vlog
  narrow interfaces, Host (CommitLSM/CommitVlog/LSMWALs/WriteMetrics +
  RequestsHaveVlogWork/UpdateHeadBuckets + throttle indicators), New,
  Start, Close, Send, plus all the moved processor/burst/sync/ack
  internals
- db.go shrinks 766 lines (2106 → 1504): the entire pipeline disappears
  from the root; db keeps a Pipeline ptr field + thin sendToWriteCh
  facade + commit.Host accessors (CommitLSM, CommitVlog, ThrottleSignal,
  UpdateHeadBuckets — UpdateHeadBuckets takes db.Lock around the
  underlying private updateHeadBuckets to absorb the locking the host
  used to do at the call site)
- CommitLSM/CommitVlog use distinct names from stats.Host's LSM/Vlog
  to avoid Go's lack-of-covariant-return-types method collision when
  one *DB satisfies both interfaces
- ApplyRequests + FinishCommitRequests exported on Pipeline so the
  root TestApplyRequests* / TestFinishCommitRequestsPerRequestErrors
  tests stay where they are without duplicating fixtures
- requestsHaveVlogWork → RequestsHaveVlogWork (now part of commit.Host
  contract)
